minor code improvements
